+++ src/Gate.IO.Api/Margin/Requests/GateMarginInterestRecordQueryRequest.cs
@@ -1,0 +1,37 @@
+namespace Gate.IO.Api.Margin;
+
+/// <summary>
+/// Isolated margin interest record query request
+/// </summary>
+public record GateMarginInterestRecordQueryRequest
+{
+    /// <summary>
+    /// Currency pair
+    /// </summary>
+    public string Symbol { get; set; }
+
+    /// <summary>
+    /// Currency
+    /// </summary>
+    public string Currency { get; set; }
+
+    /// <summary>
+    /// Page number
+    /// </summary>
+    public int? Page { get; set; }
+
+    /// <summary>
+    /// Maximum number of records returned in a single list
+    /// </summary>
+    public int? Limit { get; set; }
+
+    /// <summary>
+    /// Start timestamp
+    /// </summary>
+    public DateTime? From { get; set; }
+
+    /// <summary>
+    /// End timestamp
+    /// </summary>
+    public DateTime? To { get; set; }
+}