+++ src/Gate.IO.Api/Margin/Requests/GateMarginLoanQueryRequest.cs
@@ -1,0 +1,27 @@
+namespace Gate.IO.Api.Margin;
+
+/// <summary>
+/// Isolated margin loan query request
+/// </summary>
+public record GateMarginLoanQueryRequest
+{
+    /// <summary>
+    /// Currency pair
+    /// </summary>
+    public string Symbol { get; set; }
+
+    /// <summary>
+    /// Currency
+    /// </summary>
+    public string Currency { get; set; }
+
+    /// <summary>
+    /// Page number
+    /// </summary>
+    public int? Page { get; set; }
+
+    /// <summary>
+    /// Maximum number of records returned in a single list
+    /// </summary>
+    public int? Limit { get; set; }
+}