+++ src/Gate.IO.Api/Margin/Responses/GateMarginLoan.cs
@@ -27,17 +27,19 @@ public record GateMarginLoan
     /// Type
     /// </summary>
     [JsonProperty("type")]
-    public GateMarginUniOrderType Type { get; set; }
+    public GateMarginLoanType Type { get; set; }
 
     /// <summary>
     /// Create time
     /// </summary>
     [JsonProperty("create_time")]
+    [JsonConverter(typeof(DateTimeConverter))]
     public DateTime CreateTime { get; set; }
 
     /// <summary>
     /// Update time
     /// </summary>
     [JsonProperty("update_time")]
+    [JsonConverter(typeof(DateTimeConverter))]
     public DateTime? UpdateTime { get; set; }
 }